IOS-2282 Move blockList api to BlockService
